Modified old MlflowLogger and AbstractTrainer to allow for continue logging on a previously logged model training, invovling: - adding an optional argument for the MlflowLogger to take in a run_id, which signals the logger to resume a run. - adding additional handle for abstract trainer to allow for configuration of the starting epoch number so the newly logged metrics is associated with the correct epoch number.
These are minimal modifications needed to quickly support retraining, but still missing some key components:
- loading model and weights
- retrieving the epoch information
- matching settings

As a temp measure a helper function is added for retrieveing a single mlflow run to go hand in hand with the modified logger. Manual initialization and loading of model weights and some trainer setting is still required.